requirements.txt
torch
torchvision
matplotlib
numpy
pillow
